requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
bleach==6.0.0
Django==4.2.5
django-cors-headers==4.2.0
django-environ==0.11.2
django-filter==23.3
django-jet-reboot==1.3.3
django-summernote==0.8.20.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
importlib-metadata==6.8.0
Markdown==3.4.4
numpy==1.24.4
pandas==2.0.3
Pillow==10.0.1
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
webencodings==0.5.1
zipp==3.17.0
